+++ .nvmrc
@@ -1,1 +1,1 @@
-v24.12.0
+v24.19.0
