+++ package.json
@@ -4,7 +4,7 @@
   "version": "0.0.1",
   "private": true,
   "engines": {
-    "node": "24.12.0"
+    "node": "24.19.0"
   },
   "scripts": {
     "preinstall": "npx only-allow pnpm",
@@ -21,10 +21,11 @@
     "set-nvm": "nvm use $(cat .nvmrc)"
   },
   "dependencies": {
-    "@astrojs/check": "^0.9.6",
-    "@astrojs/react": "^4.4.2",
-    "@astrojs/rss": "^4.0.14",
-    "@astrojs/sitemap": "^3.6.0",
+    "@astrojs/check": "^0.9.10",
+    "@astrojs/markdown-remark": "^7.2.4",
+    "@astrojs/react": "^6.0.4",
+    "@astrojs/rss": "^4.0.19",
+    "@astrojs/sitemap": "^3.7.3",
     "@astrojs/tailwind": "^6.0.2",
     "@fontsource-variable/inter": "^5.0.18",
     "@fontsource-variable/noto-sans-jp": "^5.0.19",
@@ -38,7 +39,7 @@
     "@radix-ui/react-tooltip": "^1.2.8",
     "@types/react": "^18.3.1",
     "@types/react-dom": "^18.3.0",
-    "astro": "^5.16.4",
+    "astro": "^7.2.4",
     "astro-expressive-code": "^0.44.1",
     "astro-seo": "^0.8.4",
     "class-variance-authority": "^0.7.0",
@@ -59,7 +60,7 @@
     "typescript": "^5.4.5"
   },
   "devDependencies": {
-    "@astrojs/partytown": "^2.1.4",
+    "@astrojs/partytown": "^2.1.7",
     "prettier": "^3.2.5",
     "prettier-plugin-astro": "^0.13.0",
     "prettier-plugin-tailwindcss": "^0.6.11",
