--- src/content/config.ts
@@ -1,17 +1,0 @@
-import { z, defineCollection } from 'astro:content';
-
-const blogCollection = defineCollection({
-  type: 'content',
-  schema: z.object({
-    title: z.string(),
-    pubDate: z.string(),
-    description: z.string(),
-    author: z.string(),
-    tags: z.array(z.string()),
-    externalUrl: z.string().url().optional(),
-  }),
-});
-
-export const collections = {
-  'blog': blogCollection,
-};
